refactor: moved files to outside utils folder
