+++ bistrain/noise.py
