+++ bistrain/trainer.py
@@ -5,10 +5,10 @@ import gym
 import numpy as np
 from gym.core import Env
 
-from .utils.noise import OUNoise, GaussianNoise
+from .noise import OUNoise, GaussianNoise
 from .agents import DDPGAgent, A2CAgent
-from .utils.configuration import BisTrainConfiguration, LocalConfig
-from .utils.experience import make_multi_envs
+from .configuration import BisTrainConfiguration, LocalConfig
+from .utils import make_multi_envs
 
 # In case of being imported on notebook
 try:
