+++ bistrain/utils/__init__.py
@@ -1,0 +1,6 @@
+"""
+Import utilities functions
+"""
+from .utils import (make_multi_envs,
+                    soft_update,
+                    n_step_boostrap)
